% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rv.R
\name{generate_rv}
\alias{generate_rv}
\title{Random number generation for Correlated Survival and Ordinal Outcomes}
\usage{
generate_rv(
  target_tau,
  num_samples,
  censoring_rate,
  copula_type = c("Gaussian", "Clayton"),
  x_levels = 5,
  x_probs = c("equal", "random"),
  x_values = NULL,
  hr = 1
)
}
\arguments{
\item{target_tau}{the target value of the Kendall's \eqn{\tau}}

\item{num_samples}{the number of simulation samples}

\item{censoring_rate}{the censoring rate}

\item{copula_type}{copula type (default = "Gaussian")
\itemize{
\item \code{Gaussian}: Gaussian copula
\item \code{Clayton}: Clayton copula
}}

\item{x_levels}{the number of category for the ordinal outcome (default = 5)}

\item{x_probs}{the proportion of each level of x (default = "equal")
\itemize{
\item \code{equal}: equal sample sizes
\item \code{random}: random sample sizes
}}

\item{x_values}{the cutoff values of x. If \code{x_values} is specified,
\code{x_probs} is ignored. \code{x_levels - 1} and the
lengths of \code{x_levels} must be equal.}

\item{hr}{the hazard rate for survival time (default = 1.0)}
}
\value{
\itemize{
\item \code{y}: the survival time or censoring time outcome vector
\item \code{event}: the event indicator outcome vector
\item \code{x}: the ordinal (or continuous) outcome vector
\item \code{t}: the true survival time vector (for simulation)
\item \code{c}: the true censoring time vector (for simulation)
\item \code{x0}: \code{x} before discretization (for simulation)
}
}
\description{
This function generates random number vectors of correlated survival and
ordinal outcomes. Gaussian or Clayton copula is used to specify the
correlation structure. Marginal distributions are based on a exponential
distribution. The order outcome is generated by discretizing one of the
correlated random numbers based on the quantile.
}
\examples{
library(surrosurvo)
set.seed(1234)
data <- generate_rv(0.7, 500, 0.3)
surrosurvo(data$y, data$event, data$x)

# 5-level ordinal variable with Clayton copula
data <- generate_rv(0.7, 500, 0.1, copula_type = "Clayton", x_levels = 5)
surrosurvo(data$y, data$event, data$x)
}
