% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrosurvo-package.R
\docType{package}
\name{surrosurvo-package}
\alias{surrosurvo-package}
\title{Evaluation of Surrogacy for Survival and Ordinal Endpoints}
\description{
For simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nshi-stat.github.io/surrosurvo/}
}

}
\author{
Kengo Nagashima and Jun Okui
}
\keyword{internal}
\keyword{package}
