% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_srv.R
\name{generate_srv}
\alias{generate_srv}
\title{Random number generation for Correlated Survival and Survival Outcomes}
\usage{
generate_srv(
  target_tau,
  num_samples,
  censoring_rate,
  censoring_ratex = NULL,
  censtype = c("univariate", "independent"),
  copula_type = c("Gaussian", "Clayton"),
  hr = 1
)
}
\arguments{
\item{target_tau}{the target value of the Kendall's \eqn{\tau}}

\item{num_samples}{the number of simulation samples}

\item{censoring_rate}{the censoring rate for \eqn{y}}

\item{censoring_ratex}{the censoring rate for \eqn{x}}

\item{copula_type}{copula type (default = "Gaussian")
\itemize{
\item \code{Gaussian}: Gaussian copula
\item \code{Clayton}: Clayton copula
}}

\item{hr}{the hazard rate for survival time \eqn{y} (default = 1.0)}
}
\value{
\itemize{
\item \code{y}: the survival time or censoring time outcome vector \eqn{y}
\item \code{event}: the event indicator outcome vector \eqn{y}
\item \code{x}: the survival time or censoring time outcome vector \eqn{x}
\item \code{eventx}: the event indicator outcome vector \eqn{x}
\item \code{t}: the true survival time vector \eqn{y} (for simulation)
\item \code{c}: the true censoring time vector \eqn{y} (for simulation)
\item \code{tx}: the true survival time vector \eqn{x} (for simulation)
\item \code{cx}: the true censoring time vector \eqn{x} (for simulation)
}
}
\description{
This function generates random number vectors of correlated survival and
survival outcomes. Gaussian or Clayton copula is used to specify the
correlation structure. Marginal distributions are based on a exponential
distribution.
}
\examples{
library(surrosurvo)
set.seed(1234)
data <- generate_srv(0.7, 500, 0.1, 0.1)
surrosurvo(data$y, data$event, data$x, data$eventx)
}
