% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrosurvo-package.R
\docType{package}
\name{surrosurvo-package}
\alias{surrosurvo-package}
\title{Evaluation of Individual-Level Surrogacy for Survival and Ordinal,
Continuous, or Survival Endpoints in Data with Ties}
\description{
An implementation of estimators of Kendall's tau for survival data: Oakes (1982)'s non-parametric estimator \doi{10.1111/j.2517-6161.1982.tb01222.x}, Lakhal et al. (2009)'s inverse probability weighting (IPCW) estimators \doi{10.2202/1557-4679.1121}, Okui et al. (2025)'s modified IPCW estimator \doi{10.1093/bjs/znaf131}, and theses extensions of stratified estimators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nshi-stat.github.io/surrosurvo/}
}

}
\author{
Kengo Nagashima and Jun Okui
}
\keyword{internal}
\keyword{package}
