% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrosurvo-package.R
\docType{package}
\name{surrosurvo-package}
\alias{surrosurvo-package}
\title{Evaluation of Surrogacy for Survival and Ordinal Endpoints}
\description{
For simulation.
}
\author{
Kengo Nagashima, Jun Okui, , and Yasunori Sato
}
\keyword{internal}
\keyword{package}
